Implementation for pytesting unit tests on scripts
